requirements.txt
python-dotenv
fastapi 
uvicorn 
httpx 
python-multipart